Replace new cases and cases per country plot
A new plot that shows new confirmed cases and total cases with double
y-axes is added. The plot can also be switched to show number of deaths.
This plot displays data in a much clearer way that the previous ones.
It does not show data for active cases and recoveries but those are
generally not trustworthy anyway.